Administrator Profile#1 + Role-based Authorization + Multiple Log-in Prevention
Created the administrator profile, which, at the moment, shows the
database on a GridView element.

Made it so that only administrators could access the webpages destinated
to administration purposes.

Made it so the already logged-in user could not log-in again, causing a
website crash. Instead, an error message will pop up, not allowing him
to log-in while he's already logged to an account.